requirements.txt
scikit-learn 
influxdb-client
neptune-client
neptune-sklearn
flask
